fix: rename `openssl/md5.h` to avoid include conflicts with OpenSSL
Ref: #7987
